feat: Reconfigure Bind Service
Just changing the dns records do not activate them. You have to reconfigure the bind service for that. The swarm OPNSense updater does now do that. You can use the reconfigure command for that.

Adding and removing dsn records due to docker events automatically reconfigure the bind service.

The projekt build was changed from make to poetry. This is not only the python way but also comes with a modern dependency management. The former pip approach was removed from the project.

All project information, build configuration and depencies are managed using the pyproject.toml file.

The project layout was changed from flat and non standard, to the standardized model layout. This way the poetry build does understand our project and can build it wihtout any trouble.